requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
python-multipart>=0.0.9
pydantic-settings==2.2.1
email-validator>=2.0.0
python-dotenv==1.0.1
pillow==10.3.0
requests==2.31.0
easyocr==1.7.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
sentence-transformers==2.7.0
pypdf2==3.0.1
numpy==1.26.4
httpx==0.27.0
groq==0.5.0
inference-sdk==0.9.21
